features/fix_presets_and_overclocking

+++ src/Management/MNX.MonitoringCenter.Management.DataAccess/Preset/PresetRepository.cs
@@ -5,7 +5,6 @@ using MNX.MonitoringCenter.Management.UseCases;
 using MNX.MonitoringCenter.Management.Core.Overclocking;
 using MNX.MonitoringCenter.Management.DataAccess.Overclocking;
 using MNX.MonitoringCenter.Management.UseCases.Overclocking.Presets;
-using MNX.MonitoringCenter.Management.Core.Mining.MiningDevice;
 
 namespace MNX.MonitoringCenter.Management.DataAccess.Preset;
 
@@ -115,26 +114,31 @@ public class PresetRepository : IPresetRepository
     /// <inheritdoc/>
     public async Task Update(Preset preset)
     {
-        var oldOverclocking = await _context.Overclocking
-            .FirstOrDefaultAsync(x => x.Id == preset.OverclockingId);
-        if (oldOverclocking != null)
-        {
-            _context.Overclocking.Remove(oldOverclocking);
-        }
-
         var overclocking = _mapper.Map<OverclockingDto>(preset.Overclocking);
-        await _context.Overclocking.AddAsync(overclocking);
-
-        preset.OverclockingId = overclocking.Id;
+        _context.Overclocking.Update(overclocking);
         _context.Presets.Update(preset);
         await _context.SaveChangesAsync();
     }
 
     /// <inheritdoc/>
-    public Task Remove(Guid id, Guid userId)
+    public async Task Remove(Guid id, Guid userId)
     {
-        return _context.Presets
+        var removablePreset = await _context.Presets.AsNoTracking()
             .Where(x => x.Id == id && x.UserId == userId && x.IsVisible)
+            .FirstOrDefaultAsync();
+
+        if (removablePreset == null) return;
+
+        var overclockingId = removablePreset.OverclockingId;
+
+        await _context.Presets
+            .AsNoTracking()
+            .Where(x => x.Id == id && x.UserId == userId && x.IsVisible)
+            .ExecuteDeleteAsync();
+
+        await _context.Overclocking
+            .AsNoTracking()
+            .Where(x => x.Id == overclockingId)
             .ExecuteDeleteAsync();
     }
 }
